Refactor S3 storage adapter modules
