+++ arclith/adapters/outbound/s3/__init__.py
@@ -1,3 +1,4 @@
-from arclith.adapters.outbound.s3.file_storage import S3FileStorage, S3StorageConfig
+from arclith.adapters.outbound.s3.config import S3StorageConfig as S3StorageConfig
+from arclith.adapters.outbound.s3.file_storage import S3FileStorage as S3FileStorage
 
 __all__ = ["S3FileStorage", "S3StorageConfig"]
